+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Batch Changelog
 
+## 1.2.0 - 2017.08.20
+
+### Added
+* Support for batch deletion of entries & users
+* Lightswitch field appears when setting a lightswitch field value
+* Support for deleting all blocks within a matrix field
+
+### Changed
+* Now displays batch configuration in a table layout
+
+### Fixed
+* Entry type filtering
+
 ## 1.1.0 - 2017.08.18
 
 ### Added